+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.6
 ### is.codion
 - Dependency cleanup using dependency-analysis-gradle-plugin.
+### is.codion.common.core
+- ProxyBuilder.DefaultHandler.invoke() bug fixed, now unwraps the InvocationTargetException and rethrows the target exception.
 ### is.codion.framework.domain.test
 - EntityFactory.populateForeignKeys() removed, related refactoring.
 ### is.codion.framework.db.core